pyproject.toml
[tool.poetry]
name = "html-parsing"
version = "0.1.0"
description = ""
authors = ["CPR Tech <tech@climatepolicyradar.org>"]

[tool.poetry.dependencies]
python = "^3.9"
awscli = "1.25.69"
news-please = "^1.5.22"
pandas = "^1.4.4"
tqdm = "^4.64.1"
pydantic = "^1.10.2"
click = "^8.1.3"
langdetect = "^1.0.9"
playwright = "^1.25.2"
readability-lxml = "^0.8.1"
bleach = "^5.0.1"
python-json-logger = "^2.0.4"
torchvision = "^0.13.1"
torch = "^1.12.1"
layoutparser = {extras = ["ocr"], version = "^0.3.4"}
Shapely = "^1.8.4"
cloudpathlib = {extras = ["s3"], version = "^0.10.0"}
PyMuPDF = "^1.20.2"
google-cloud-translate = "^3.8.2"
psutil = "^5.9.2"
multiprocessing-logging = "^0.3.3"
json-logging = "^1.3.0"
tenacity = "^8.2.1"
google-cloud-documentai = "^2.16.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pyright = "^1.1.270"
pytest = "^7.1.3"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pydocstyle]
# D100 - Missing docstring in public module
# D103 - Missing docstring in public function
# D104 - Missing docstring in public package
# D107 - Missing docstring in __init__
# D202 - No blank lines allowed after function docstring
# D203 - 1 blank line required before class docstring
# D212 - Multi-line docstring summary should start at the first line
# D213 - Multi-line docstring summary should start at the second line
# D400 - First line should end with a period
# D401 - First line should be in imperative mood (Google styleguide denotes 'descriptive' style.)
# D406 - Section name should end with a newline (incompatible with Google styleguide)
# D407 - Missing dashed underline after section (incompatible with Google styleguide)
# D413 - Missing blank line after last section (incompatible with Google styleguide)
# D415 - First line should end with a period, question mark, or exclamation point
ignore = """
D100,D103,D104,D107,
D202,D203,D212,D213,
D400,D401,D406,D407,D413,D415
"""

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"

[tool.black]
line-length = 88
target-version = ['py39']
